move ElectionRules class into its own file
