+++ backend/tests/electionRules.py
@@ -1,7 +1,7 @@
 # coding:utf-8
 from unittest import TestCase
 
-from voting import ElectionRules
+from election_rules import ElectionRules
 from simulate import generate_opt_ruleset
 from simulate import generate_law_ruleset
 from simulate import generate_ind_const_ruleset
